+++ src/TestSuite/EntityTestCase.php
@@ -57,12 +57,13 @@ abstract class EntityTestCase extends TestCase
      * Asserts that the entity has a "no accessible" property
      * @param string|array $property Property name
      * @return void
-     * @todo missing params
      */
-    public function assertHasNoAccessibleProperty($property): void
+    public function assertHasNoAccessibleProperty($property, string $message = ''): void
     {
+
         foreach ((array)$property as $name) {
-            $this->assertFalse($this->Entity->isAccessible($name));
+            $message = $message ?: sprintf('Failed asserting that `\%s::$%s` property is not accessible', get_class($this->Entity), $name);
+            $this->assertFalse($this->Entity->isAccessible($name), $message);
         }
     }
 }