+++ tests/TestCase/TestSuite/EntityTestCaseTest.php
@@ -18,6 +18,7 @@ namespace MeCms\Test\TestCase\TestSuite;
 use MeCms\Model\Entity\Post;
 use MeCms\Test\TestCase\Model\Entity\PostTest;
 use MeCms\TestSuite\TestCase;
+use PHPUnit\Framework\AssertionFailedError;
 
 /**
  * EntityTestCaseTest class
@@ -26,7 +27,7 @@ class EntityTestCaseTest extends TestCase
 {
     /**
      * @test
-     * @uses \MeCms\TestSuite\CellTestCase::__get()
+     * @uses \MeCms\TestSuite\EntityTestCase::__get()
      */
     public function testMagicGetMethod(): void
     {
@@ -34,5 +35,20 @@ class EntityTestCaseTest extends TestCase
         $this->assertSame('Posts', $EntityTestClass->alias);
         $this->assertSame(Post::class, $EntityTestClass->originClassName);
         $this->assertInstanceOf(Post::class, $EntityTestClass->Entity);
+    }
+
+    /**
+     * @test
+     * @uses \MeCms\TestSuite\EntityTestCase::assertHasNoAccessibleProperty()
+     */
+    public function testAssertHasNoAccessibleProperty(): void
+    {
+        $EntityTestClass = new PostTest();
+        $EntityTestClass->assertHasNoAccessibleProperty('preview');
+        $EntityTestClass->assertHasNoAccessibleProperty(['preview', 'modified']);
+
+        $this->expectException(AssertionFailedError::class);
+        $this->expectExceptionMessage('Failed asserting that `\MeCms\Model\Entity\Post::$preview` property is not accessible');
+        $EntityTestClass->assertHasNoAccessibleProperty(['preview', 'title']);
     }
 }